Fixed invoices
now we only has 1 invoice

+++ ESD6NL.DriverSystem.DAL/Interfaces/IInvoiceRepository.cs
@@ -7,6 +7,6 @@ namespace ESD6NL.DriverSystem.DAL.Interfaces
 {
     public interface IInvoiceRepository : IGenericRepository<Invoice>
     {
-        Invoice GetSpecificInvoice(int invoiceId);
+        Invoice GetSpecificInvoice(long invoiceId);
     }
 }
